basic-python/requirements.txt
netmiko
ntc-templates
flask
pyyaml
